homework10-web/back/requirements.txt
psycopg2
pg8000
sqlalchemy
postgres
fastapi
uvicorn
passlib
python-jose
python-multipart
bcrypt